feat: Add a deployment stack with clickhouse
Migrate from Elasticsearch and Kibana to Clickhouse. To make it easier accessing clickhouse data, the clickhouse web ui
is added in a dedicated container. It requires to have access from the client browser to clickhouse and requires to set
the Clickhouse host host or FQDN by the user. Add and configure Grafana with ClickHouse plugin and configuration
provisioning.

Signed-off-by: Ronny Trommer <ronny@no42.org>