feat: Add Riptide release target
Create a Makefile target helping with releases. The release target takes care of setting version number in the pom.xml
for the version to be released, tagging a release and setting a new SNAPSHOT version in main. By default the release
isn't pushed to GitHub and runs just locally. The release engineer needs to push the tags and commits to the upstream
repository.

Signed-off-by: Ronny Trommer <ronny@no42.org>